requirements.txt
fastapi
uvicorn
faiss-cpu
openai
pydantic
python-dotenv
python-multipart
python-pptx
PyMuPDF